[TS migration] Migrate 'AnonymousReportFooter.js' component
